leetcode-30day-challenge/src/main/java/dev/wonho/leetcode/week2/BackspaceStringCompare.java
package dev.wonho.leetcode.week2;

import java.util.Stack;

/*
Day 9 - Backspace String Compare

Given two strings S and T, return if they are equal when both are typed into empty text editors.
# means a backspace character.

Example 1:
    Input: S = "ab#c", T = "ad#c"
    Output: true
    Explanation: Both S and T become "ac".

Example 2:
    Input: S = "ab##", T = "c#d#"
    Output: true
    Explanation: Both S and T become "".

Example 3:
    Input: S = "a##c", T = "#a#c"
    Output: true
    Explanation: Both S and T become "c".

Example 4:
    Input: S = "a#c", T = "b"
    Output: false
    Explanation: S becomes "c" while T becomes "b".

Note:
    1 <= S.length <= 200
    1 <= T.length <= 200
    S and T only contain lowercase letters and '#' characters.

Follow up:
    Can you solve it in O(N) time and O(1) space?
 */
public class BackspaceStringCompare {

    public boolean backspaceCompare(String S, String T) {
        return applyBackspace(S).equals(applyBackspace(T));
    }

    private String applyBackspace(String str) {
        Stack<Character> stack = new Stack<>();

        for (int i = 0; i < str.length(); i++) {
            if (str.charAt(i) != '#') {
                stack.push(str.charAt(i));
            } else if (!stack.empty()) {
                stack.pop();
            }
        }

        return stack.toString();
    }

}
